+++ CHANGELOG.md
@@ -90,6 +90,7 @@
 - PR #2003 Removed few redundant unit-tests from test_string.py::test_string_cat
 - PR #1944 Groupby design improvements
 - PR #2017 Convert `read_orc()` into a C++ API
+- PR #2011 Convert `read_parquet()` into a C++ API
 - PR #1756 Add documentation "10 Minutes to cuDF and dask_cuDF"
 - PR #2034 Adding support for string columns concatenation using "add" binary operator
 - PR #2042 Replace old "10 Minutes" guide with new guide for docs build process